+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - Fix structure tab not shown [#369](https://github.com/etalab/udata-front/pull/369)
 - Update sentry configuration to ignore aborted requests [#368](https://github.com/etalab/udata-front/pull/368)
 - Fix CI not using udata release when needed [#363](https://github.com/etalab/udata-front/pull/363)
+- Use udata schema endpoint [#336](https://github.com/etalab/udata-front/pull/336)
 
 ## 3.5.3 (2024-02-22)